Revise the makefile and create-all-folder scripts to create and copy PNGs in various sizes
This commit heavily refactors the create-all-folder script to use functions to reduce
verbosity and make it easier to add new folders in the future.  It also copies the PNGs
in all the available sizes to the all folder.

The build-pngs target in the Makefile will now generate PNGs in 16, 32, 64 and 128 pixel
sizes