Icon Generation script
This commit gets the icon generation script making icons that don't look awful.  The initial
draft used convert on the SVG but when used in that mode it resizes awfully.  The new version
using the existing PNGs to generate the icon since the PNGs should be nicely sized.

This also expands the sizes of PNGs generated by the make file to include more sizes that are
typically included in ICO files.

I encountered some slightly weird artifacts in the 128x128 icons but not sure if this is simply
an artifact of the slightly borked ImageMagick install on my OS X box.