Improve how readme is generated to use template files which can be easily edited outside the script
This moves the header and main text of the README into separate .md files which are added to
the README generated by update-readme.sh using cat.  This makes it much easier to edit the
text separate from the script.

I also improved the text to discuss how to generate PNGs in different sizes and to cover
the newly added ICO support